Перенес генерацию моков в папку utils
